requirements.txt
pydantic~=2.10.6
openai~=1.66.3
tenacity~=9.0.0
pyyaml~=6.0.2
loguru~=0.7.3
structlog~=25.4.0
numpy
datasets~=3.4.1
fastapi~=0.115.11
tiktoken~=0.9.0

html2text~=2024.2.26
gymnasium~=1.1.1
pillow~=10.4.0
browsergym~=0.13.3
uvicorn~=0.34.0
unidiff~=0.7.5
browser-use~=0.1.40
googlesearch-python~=1.3.0
baidusearch~=1.0.3
duckduckgo_search~=7.5.3

aiofiles~=24.1.0
pydantic_core~=2.27.2
colorama~=0.4.6
playwright~=1.51.0

docker~=7.1.0
daytona==0.21.8
pytest~=8.3.5
pytest-asyncio~=0.25.3

mcp~=1.5.0
httpx>=0.27.0
lark-oapi~=1.5.3
tomli>=2.0.0

boto3~=1.34.106

requests~=2.32.3
beautifulsoup4~=4.13.3
crawl4ai~=0.6.3

huggingface-hub~=0.29.2
setuptools~=75.8.0
qdrant-client~=1.12.1
chromadb~=0.5.23
sentence-transformers~=3.4.1
rank-bm25~=0.2.2
pypdf~=5.3.1
python-docx~=1.1.2
unstructured~=0.16.23
python-multipart~=0.0.20
